requirements.txt
Flask==2.3.3
Flask-Session==0.5.0
mysql-connector-python==8.3.0
Werkzeug==2.3.7
pytz